fix
Former-commit-id: 2c99d061aee444985c2c2082d0e4c2fd2f05b2ff

+++ client/cmain.go
@@ -65,7 +65,7 @@ func parse(str string) interface{} {
 		}
 	case "DeleteCharacter":
 		i, _ := strconv.Atoi(args[1])
-		return &client.DeleteCharacter{CharacterIndex: int16(i)}
+		return &client.DeleteCharacter{CharacterIndex: int32(i)}
 	case "StartGame", "start", "startgame": // 8
 		i, _ := strconv.Atoi(args[1])
 		return &client.StartGame{CharacterIndex: int16(i)}
